+++ core/src/main/java/org/vertexium/GraphWithSearchIndex.java
@@ -4,4 +4,9 @@ import org.vertexium.search.SearchIndex;
 
 public interface GraphWithSearchIndex extends Graph {
     SearchIndex getSearchIndex();
+
+    /**
+     * This method will only flush the primary graph and not the search index
+     */
+    void flushGraph();
 }